fixxx events

+++ events/views.py
@@ -77,14 +77,15 @@ def edit_event(request, id):
         form = EventForm(request.POST, request.FILES, instance=event)
         if form.is_valid():
             form.save()
-            return redirect('events:show_events')
+            return redirect('events:event_list')
     else:
         form = EventForm(instance=event)
-    return render(request, 'edit_event.html', {'form': form, 'event': event})
+    return render(request, 'events/edit_event.html', {'form': form, 'event': event})
 
 @login_required(login_url='/login/')
 def delete_event(request, id):
     event = get_object_or_404(Event, id=id)
-    if request.user.is_staff:  # hanya admin/staff
+    if request.user.is_staff and request.method == 'POST':
         event.delete()
-    return redirect('events:show_events')
+    return redirect('events:event_list')
+
